+++ src/short-urls/data/index.ts
@@ -1,3 +1,4 @@
+import type { ShlinkVisitsSummary } from '../../api/types';
 import type { Order } from '../../utils/helpers/ordering';
 import type { Nullable, OptionalString } from '../../utils/utils';
 
@@ -41,7 +42,7 @@ export interface ShortUrl {
   dateCreated: string;
   /** @deprecated */
   visitsCount: number; // Deprecated since Shlink 3.4.0
-  visitsSummary?: ShortUrlVisitsSummary; // Optional only before Shlink 3.4.0
+  visitsSummary?: ShlinkVisitsSummary; // Optional only before Shlink 3.4.0
   meta: Required<Nullable<ShortUrlMeta>>;
   tags: string[];
   domain: string | null;
@@ -54,12 +55,6 @@ export interface ShortUrlMeta {
   validSince?: string;
   validUntil?: string;
   maxVisits?: number;
-}
-
-export interface ShortUrlVisitsSummary {
-  total: number;
-  nonBots: number;
-  bots: number;
 }
 
 export interface ShortUrlModalProps {